refactor(frontend): split tab-ai-llm into feature modules
The AI & LLM settings tab lived in a single ~2200-line module bundling four
unrelated feature areas — LLM connection settings, MCP server management, the
prompt-profiles feature, and advanced AI tuning — which made it a merge-conflict
magnet and prevented Vite from code-splitting the admin-only prompt/tuning UI off
non-admin paths.

Mechanically split the module into a tab-ai-llm/ directory of cohesive files with
no logic changes:
- tab-ai-llm.tsx: AiLlmTab, LlmSettingsSection, AdvancedAiTuningSection
- mcp-servers.tsx: McpServersSection, McpServerRow
- prompts.tsx: AiPromptsTab, ProfileSelector, PromptTestPanel, ImportPreviewPanel,
  PromptHistoryPanel and their local helpers (TokenBadge, estimateTokens,
  computeDiff, formatRelativeTime)
- index.ts: re-exports the public API so the original import path
  (@/features/core/components/settings/tab-ai-llm) keeps working unchanged for
  settings.tsx and the existing tests.

Every exported name, data-testid, and behaviour is identical; component bodies were
moved verbatim (preserving the earlier accessibility and shared relative-time
changes). No callers or tests were modified.

Closes #1513
