+++ frontend/src/features/core/components/settings/tab-ai-llm/index.ts
@@ -1,0 +1,6 @@
+// Barrel for the AI & LLM settings tab, split from a single ~2200-line module into
+// cohesive feature files. Re-exports preserve the original import path
+// (`@/features/core/components/settings/tab-ai-llm`) for callers and tests.
+export { AiLlmTab, LlmSettingsSection, LLM_SETTING_KEYS } from './tab-ai-llm';
+export { McpServerRow, McpServersSection } from './mcp-servers';
+export { AiPromptsTab, PromptTestPanel, ImportPreviewPanel } from './prompts';
